sectioned off the windowMeasure, clsOrient, and displayTable functions into their own file (display.py) as it will be undergoing lots of additions and I don't want to clutter up the main file
